Fixed beta dual handcannon bugs
If you have a HC, you can pickup a 2nd one regardless of how much weight
you are carrying.  The Beta HC would add 4 blocks to the weight check,
tus if you have 12/15 blocks, you cannot pick up a 2nd one.

The second bug was in the trader.  When you sell the Beta HC in the trader,
it would give you the default HC.